// typescript/user-interface/packages/ui-state/src/ts/app/query/types.ts
interface BaseAsyncFetchRequest<T> {
  /** the number of pending (active) fetch requests */
  pending: number;
  /** the number of fulfilled (completed) fetch requests */
  fulfilled: number;
  /** the number of rejected fetch requests */
  rejected: number;
  /** true if fetching is active or loading; false otherwise */
  isLoading: boolean;
  /** true if an error occurred while fetching; false otherwise */
  isError: boolean;
  /** the data of the result */
  data?: T;
  /** the serializable errors */
  error?: (SerializedError | SerializedAxiosError)[];
}

/**
 * Defines the base result object for an async fetch request.
 */
export type AsyncFetchResult<T> =
  | Override<
      BaseAsyncFetchRequest<T>,
      | Omit<BaseAsyncFetchRequest<T>, 'error'>
      | ({
          isLoading: false;
          isError: false;
        } & Required<Pick<BaseAsyncFetchRequest<T>, 'data'>>)
    >
  | Override<
      BaseAsyncFetchRequest<T>,
      | Omit<BaseAsyncFetchRequest<T>, 'data'>
      | {
          isLoading: true;
          isError: false;
        }
    >
  | Override<
      BaseAsyncFetchRequest<T>,
      | Omit<BaseAsyncFetchRequest<T>, 'data'>
      | ({
          isLoading: false;
          isError: true;
        } & Required<Pick<BaseAsyncFetchRequest<T>, 'error'>>)
    >;
// ...
export type FetchHistoryStatus = Omit<AsyncFetchResult<never>, 'data'>;